Update TheCipherGame.java
I removed the boolean field which was used to determine whether a message has been encoded using the AIZ26 cipher, and I replaced it with a regex expression. This is a vast improvement because, previously, a user could not go directly from an encrypted AIZ26 message to plain text (due to the fact that the boolean mentioned earlier was automatically set to false).

In addition, I fixed the formatedKey function, which is used for encoding and decoding Vigenere messages. It is now possible to have cipher keys and encompass many words, instead of the usual one.